22/10/2024 pull request

+++ CondTools/SiPixel/test/SiPixelDynamicInefficiencyDB_cfg.py
@@ -1,4 +1,5 @@
-#import os
+from __future__ import print_function
+import os
 import shlex, shutil, getpass
 #import subprocess
 
@@ -6,21 +7,19 @@ import FWCore.ParameterSet.Config as cms
 
 process = cms.Process("SiPixelInclusiveBuilder")
 process.load("FWCore.MessageService.MessageLogger_cfi")
-#process.MessageLogger.destinations = cms.untracked.vstring("cout")
-#process.MessageLogger.cout = cms.untracked.PSet(threshold = cms.untracked.string("INFO"))
+#process.MessageLogger.cerr.enable = False
+#process.MessageLogger.cout = dict(threshold = "INFO", enable = True)
 
 process.load("Configuration.StandardSequences.MagneticField_cff")
 
 #hptopo
-process.load("Configuration.StandardSequences.FrontierConditions_GlobalTag_condDBv2_cff")
-from Configuration.AlCa.autoCond_condDBv2 import autoCond
+process.load("Configuration.StandardSequences.FrontierConditions_GlobalTag_cff")
+from Configuration.AlCa.autoCond import autoCond
 process.GlobalTag.globaltag = autoCond['run2_design']
-print process.GlobalTag.globaltag
+print(process.GlobalTag.globaltag)
 process.load("Configuration.StandardSequences.GeometryDB_cff")
 
 process.load("CondTools.SiPixel.SiPixelGainCalibrationService_cfi")
-
-process.load("CondCore.DBCommon.CondDBCommon_cfi")
 
 process.source = cms.Source("EmptyIOVSource",
     firstValue = cms.uint64(1),
@@ -45,11 +44,12 @@ user = getpass.getuser()
 #file = "/tmp/" + user + "/SiPixelDynamicInefficiency.db"
 file = "siPixelDynamicInefficiency.db"
 sqlfile = "sqlite_file:" + file
-print '\n-> Uploading as user %s into file %s, i.e. %s\n' % (user, file, sqlfile)
+print('\n-> Uploading as user %s into file %s, i.e. %s\n' % (user, file, sqlfile))
 
 
 #standard python libraries instead of spawn processes
-shutil.move("siPixelDynamicInefficiency.db", "siPixelDynamicInefficiency_old.db")
+if(os.path.isfile('./'+file)):
+    shutil.move("siPixelDynamicInefficiency.db", "siPixelDynamicInefficiency_old.db")
 #subprocess.call(["/bin/cp", "siPixelDynamicInefficiency.db", file])
 #subprocess.call(["/bin/mv", "siPixelDynamicInefficiency.db", "siPixelDynamicInefficiency.db"])
